ChatMessage usage improvements
ChatMessage will always hold API representation of chat.
Basically the mimeType and content of ChatMessage will
hold the API version and the conversion to corresponding
network type will be done only at the time of sending.
Similarly the incoming payloads are immediately parsed
to populate the ChatMessage with API representation of
content and mimetype.
By this, redundant conversions at many places is removed.

Change-Id: Ib45a6b683eafb9086aae30030ff10824fa0b5cc4